Point_5_Automatisation/streaming/requirements.txt
pymongo
websocket-client
python-dotenv
pandas